rename : authentication -> auth - reason : for unification of name (authentication)
